frontend/kyogre_pdv_app/app/AI_agentes_System/AI_System_Studio/requirements.txt
fastapi
uvicorn
google.generativeai
python-dotenv